added Astroid and Laser functionality

+++ src/mennospijker/icasteroids/RedLaser.java
@@ -7,7 +7,7 @@ import java.util.List;
 /**
  * The type Laser.
  */
-public class Laser extends Weapon {
+public class RedLaser extends Weapon {
     private ICAstroids world;
     private Player player;
 
@@ -18,11 +18,11 @@ public class Laser extends Weapon {
      * @param x     the x
      * @param y     the y
      */
-    public Laser(ICAstroids world, Player player, int x, int y){
+    RedLaser(ICAstroids world, Player player, int x, int y){
         super(world.loadSprite("redLaserBeam.png"), 1);
         setX(x);
         setY(y);
-
+        this.player = player;
         this.world = world;
         setDirectionSpeed(0, 5);
     }
@@ -30,18 +30,24 @@ public class Laser extends Weapon {
     @Override
     public void gameObjectCollisionOccurred(List<GameObject> list) {
         for (GameObject g : list){
-            if (g instanceof Asteroid){
+            if (g instanceof Asteroid) {
+                Asteroid asteroid = (Asteroid) g;
 
                 float x = this.getCenterX();
                 float y = this.getCenterY();
 
-                world.deleteGameObject(g);
                 world.deleteGameObject(this);
+                asteroid.hit();
+
+                if (asteroid.isDestroyed()) {
+                    player.addPoints(asteroid.getValue());
+                    asteroid.explode();
+                    world.deleteGameObject(g);
+                    player.removeLaserBeam(this);
+                }
+                world.setDifficulty();
+
             }
         }
-    }
-
-    @Override
-    public void update() {
     }
 }
